requirements.txt
# Standard libraries for API requests and environment variables
requests
python-dotenv
